+++ CHANGELOG.md
@@ -11,6 +11,7 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
  * `master` branch is now called `main`.
  * Rename `Grok::with_patterns()` to `Grok::with_default_patterns()`.
  * Rename `Grok::insert_definition` to `Grok::add_pattern` to stop mixing "definition" with "pattern".
+ * Add `IntoIter` for `&Matches` for more convenient match iteration (i.e. for loop).
 
 ## [1.2.0] - 2021-03-21